rm: combine menu item
